fix: rename BoardSwitch to BoardMenu
